Only consider initialized devices
It turns out that udev nodes have a warm-up like state where
they are present in the system, but not actually ready to be
used. This is an issue for at least the C270 cameras, which
throw errors if used while in this state.

It is hoped that this change will resolve some of the
intermittent connection issues we've seen on the other
boards too.